// client/src/app/models/product.model.ts
export interface Product {
  id: number;
  name: string;
  calories_per_100g: number;
  protein_per_100g: number;
  fat_per_100g: number;
  carbs_per_100g: number;
  created_at: string;
}

export interface ProductCreateRequest {
  name: string;
  calories_per_100g: number;
  protein_per_100g: number;
  fat_per_100g: number;
  carbs_per_100g: number;
}